Add files via upload
This is my template website so far. Everything can be changed (colors, tabs, layout etc). There is a spot for a logo if we want one, I included a login button so we could link that. I did all the css inside the html file. If we wanted to we can make a seperate css file, but for simplicity I did it this way. This was done in vscode.